+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Add actor ID to unsolicited message report.
+
 ## [2.7.1] - 2024-01-13
 
 ### Changed